don't speed up `fastloop()` under accelerated time
Instead of simply running the game in tighter loops, multiply resource
calculations by game speed. This is only done for `fastloop()` for now
as most other code is tightly build and balanced around the timings of
`midloop()` and `longloop()`.

In a few places a `++` was replaced by `4*time_multiplier` to account
for the loop running 4 times per second and `time_multiplier` being
0.25 under normal game speed.