adjust timers under accelerated time
The previous versions all assumed `midloop()` would run exactly once a
second.  This is not the case under accelerated time.  This will
correctly subtract the time between `midloop()` calls from timers such
that the displayed timer accurately represents real time left.

Currently this is done for the spy's training and mission timers as well
as the gene mutation timer.