fix arpa project times in build queue
All queued items' times are based on the remaining segments of the
currently being build project, making the time for subsequent queued
items of the same project useless. This will calculate the correct
times. Do this by keeping track of which projects are queued up and only
calculate time based on remaining segments for the first item in the
queue of each project, otherwise calculate time based on building 100%
of segments.